feat(monorepo): New file export organization for components and types
