+++ packages/xprog-prensa/index.ts
@@ -1,7 +1,4 @@
-export * from './primitives/Block'
-export * from './primitives/Button'
-export * from './primitives/IconButton'
-export * from './primitives/Typography'
+export * from './primitives'
 export * from './news'
 export * from './providers/PrensaThemeProvider'
 export * from './prensa.default.config'
