+++ packages/xprog-prensa/primitives/index.ts
@@ -1,0 +1,5 @@
+export * from './Accordion/index'
+export * from './Block/index'
+export * from './Button/index'
+export * from './IconButton/index'
+export * from './Typography/index'
