+++ packages/xprog-prensa/primitives/types.ts
@@ -1,0 +1,5 @@
+export * from './Accordion/types'
+export * from './Block/props'
+export * from './Button/props'
+export * from './IconButton/props'
+export * from './Typography/props'
